+++ programs/marginfi/src/lib.rs
@@ -218,16 +218,27 @@ pub mod marginfi {
 
     // User instructions
 
-    /// Initialize a marginfi account for a given group
+    /// Initialize a marginfi account for a given group. The account is a fresh keypair, and must
+    /// sign. If you are a CPI caller, consider using `marginfi_account_initialize_pda` instead, or
+    /// create the account manually and use `transfer_to_new_account` to gift it to the owner you
+    /// wish.
     pub fn marginfi_account_initialize(ctx: Context<MarginfiAccountInitialize>) -> MarginfiResult {
         marginfi_account::initialize_account(ctx)
     }
 
-    /// Initialize a marginfi account for a given group
+    /// The same as `marginfi_account_initialize`, except the created marginfi account uses a PDA
+    /// (Program Derived Address)
+    ///
+    /// seeds:
+    /// - marginfi_group
+    /// - authority: The account authority (owner)  
+    /// - account_index: A u16 value to allow multiple accounts per authority
+    /// - third_party_id: Optional u16 for third-party tagging. Seeds < PDA_FREE_THRESHOLD can be
+    ///   used freely. For a dedicated seed used by just your program (via CPI), contact us.
     pub fn marginfi_account_initialize_pda(
         ctx: Context<MarginfiAccountInitializePda>,
-        account_index: u32,
-        third_party_id: Option<u32>,
+        account_index: u16,
+        third_party_id: Option<u16>,
     ) -> MarginfiResult {
         marginfi_account::initialize_account_pda(ctx, account_index, third_party_id)
     }
@@ -356,10 +367,18 @@ pub mod marginfi {
         marginfi_account::transfer_to_new_account(ctx)
     }
 
+    /// Same as `transfer_to_new_account` except the resulting account is a PDA
+    ///
+    /// seeds:
+    /// - marginfi_group
+    /// - authority: The account authority (owner)  
+    /// - account_index: A u32 value to allow multiple accounts per authority
+    /// - third_party_id: Optional u32 for third-party tagging. Seeds < PDA_FREE_THRESHOLD can be
+    ///   used freely. For a dedicated seed used by just your program (via CPI), contact us.
     pub fn transfer_to_new_account_pda(
         ctx: Context<TransferToNewAccountPda>,
-        account_index: u32,
-        third_party_id: Option<u32>,
+        account_index: u16,
+        third_party_id: Option<u16>,
     ) -> MarginfiResult {
         marginfi_account::transfer_to_new_account_pda(ctx, account_index, third_party_id)
     }
